перенос фикстур в тест
