+++ conftest.py
@@ -1,6 +1,3 @@
-import pytest
-
-
 def pytest_addoption(parser):
     parser.addoption(
         "--url",
@@ -14,12 +11,3 @@ def pytest_addoption(parser):
         help="expected status code",
         type=int
     )
-
-
-@pytest.fixture
-def base_url(request):
-    return request.config.getoption("--url")
-
-@pytest.fixture
-def expected_status(request):
-    return request.config.getoption("--status_code")
