+++ tests/test_expected_status_code.py
@@ -1,4 +1,15 @@
 import requests
+import pytest
+
+
+@pytest.fixture
+def base_url(request):
+    return request.config.getoption("--url")
+
+
+@pytest.fixture
+def expected_status(request):
+    return request.config.getoption("--status_code")
 
 
 def test_url_status(base_url, expected_status):
